django-server/requirements.txt
appnope==0.1.0
backcall==0.1.0
certifi==2018.4.16
chardet==3.0.4
dataclasses==0.5
decorator==4.2.1
Django==2.0.4
django-ipware==2.1.0
djangorestframework==3.7.7
djongo==1.2.26
flake8==3.5.0
idna==2.6
ipython==6.3.0
ipython-genutils==0.2.0
jedi==0.11.1
lazy-object-proxy==1.3.1
mccabe==0.6.1
mongoengine==0.15.0
newrelic==4.2.0.100
parso==0.1.1
pexpect==4.4.0
pickleshare==0.7.4
Pillow==5.1.0
prompt-toolkit==1.0.15
ptyprocess==0.5.2
pycodestyle==2.3.1
pyflakes==1.6.0
Pygments==2.2.0
pymongo==3.6.1
python-memcached==1.59
pytz==2018.3
requests==2.18.4
simplegeneric==0.8.1
six==1.11.0
sqlparse==0.2.4
traitlets==4.3.2
urllib3==1.22
uWSGI==2.0.17
wcwidth==0.1.7
wrapt==1.10.11
